feat: latest changes + fix codespell
