+++ doc/source/mapdl_commands/aux2/index.rst
@@ -6,13 +6,13 @@ Aux2
 
 .. list-table::
 
-   * - :ref:`ref_binary_file_manipulation`
    * - :ref:`ref_binary_file_dump`
+   * - :ref:`ref_binary_file_manipulation`
 
 
 .. toctree::
    :maxdepth: 1
    :hidden:
 
-   binary_file_manipulation
    binary_file_dump
+   binary_file_manipulation
